Fix wasm build regression
My last PR added a transitive dependency on pcb-kicad to pcb-zen-wasm,
which caused the wasm build to fail.

Build wasm bundle in CI to prevent future regressions.
